Most shepherd options are integrated into Report panel
The shepherd panel is constructed programmatically. However, it was
inserted into the Report tab through NetBeans visual editor. For this
the NB project first needs to be compiled. Then right click the
PtmshepherdPanel class in the Project tab, 'tools -> add to palette'.
Then you can drag and drop it from the pallete, like other components.